feat(tools,cli): expose list_agents MCP tool and pin service install paths
Add list_agents as a new MCP tool (#118) in the Identity cluster that
returns agent name, program, model, task description, inception_ts,
last_active_ts, and contact_policy for all agents in a project. Includes
TOOL_CLUSTER_MAP, metrics TOOL_META, tool directory entry, and server
registration.

Fix service install (#44) to pin absolute paths in generated systemd and
launchd units. The generated units now include WorkingDirectory and
Environment directives for DATABASE_URL and STORAGE_ROOT, resolved from
the current runtime config. Relative database URLs (the default
sqlite+aiosqlite:///./storage.sqlite3) are converted to absolute paths.

Both features include tests:
- list_agents tool: registers agents and verifies tool output
- service install: unit content tests, database URL resolution tests,
  launchd plist absolute path tests

+++ crates/mcp-agent-mail-core/src/atc_baseline.rs
@@ -501,7 +501,7 @@ mod tests {
             t.tick_budget_micros <= 50_000,
             "tick budget should be <50ms"
         );
-        assert!(t.probe_interval_micros > t.tick_budget_micros as i64);
+        assert!(t.probe_interval_micros > t.tick_budget_micros);
         assert!(t.estimated_probe_cost_micros < t.tick_budget_micros);
     }
